refactor: extract getCliDisplayName to shared util
