+++ src/core/command-handler.ts
@@ -11,7 +11,8 @@ import * as sessionStore from '../services/session-store.js';
 import * as scheduleStore from '../services/schedule-store.js';
 import * as scheduler from './scheduler.js';
 import { scanProjects, scanMultipleProjects } from '../services/project-scanner.js';
-import { buildRepoSelectCard, getCliDisplayName } from '../im/lark/card-builder.js';
+import { buildRepoSelectCard } from '../im/lark/card-builder.js';
+import { getCliDisplayName } from '../utils/cli-display.js';
 import { deleteMessage } from '../im/lark/client.js';
 import { logger } from '../utils/logger.js';
 import { getSessionCost, formatNumber } from './cost-calculator.js';
