+++ src/core/worker-pool.ts
@@ -10,7 +10,8 @@ import { randomBytes } from 'node:crypto';
 import { config } from '../config.js';
 import * as sessionStore from '../services/session-store.js';
 import { updateMessage, MessageWithdrawnError } from '../im/lark/client.js';
-import { buildStreamingCard, buildSessionCard, getCliDisplayName } from '../im/lark/card-builder.js';
+import { buildStreamingCard, buildSessionCard } from '../im/lark/card-builder.js';
+import { getCliDisplayName } from '../utils/cli-display.js';
 import { logger } from '../utils/logger.js';
 import { createCliAdapterSync } from '../adapters/cli/registry.js';
 import { TmuxBackend } from '../adapters/backend/tmux-backend.js';
